5 step done

+++ app/Http/Controllers/TaskController.php
@@ -3,10 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Task;
-use App\Models\User;
+// use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use App\Controllers\Requests\TaskValidate;
 use App\Http\Requests\StoreTask;
 
 class TaskController extends Controller
@@ -43,7 +42,7 @@ class TaskController extends Controller
         $tasks = new Task();
         $taskStatuses = DB::table('task_statuses')->get();
         $users = DB::table('users')->get();
-        $labels = [];
+        $labels = DB::table('labels')->get();
         return view('tasks.create', compact('tasks', 'taskStatuses', 'users', 'labels'));
     }
 
@@ -55,11 +54,19 @@ class TaskController extends Controller
      */
     public function store(StoreTask $request)
     {
+        $data = $request->input();
         $newTask = new Task();
         $validated = $request->validated();
         $validated['created_by_id'] = \Auth::user()->id;
         $newTask->fill($validated);
         $newTask->save();
+        $newLabels = $data['labels'];
+        if ($newLabels[0] == null) {
+            unset($newLabels[0]);
+        }
+        if (count($newLabels) > 0) {
+            $newTask->labels()->attach($newLabels);
+        }
         flash(__('flash.tasks.cteate.success'))->success();
         return redirect()->route('tasks.index');
     }
@@ -72,7 +79,11 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
-        return view('tasks.show', compact('task'));
+        $labels = [];
+        foreach ($task->labels as $labelName) {
+            $labels[] = $labelName->name;
+        }
+        return view('tasks.show', compact('task', 'labels'));
     }
 
     /**
@@ -85,16 +96,22 @@ class TaskController extends Controller
     {
         if (!\Auth::user()) {
             flash(__('flash.tasks.this_action_is_unauthorized'))->error();
-            return view('tasks.show', compact('task'));
+            return redirect()->route('tasks.show', compact('task'));
         }
         $taskStatuses = DB::table('task_statuses')->get();
         $users = DB::table('users')->get();
+        $labels = DB::table('labels')->get();
         $taskCheck = Task::find($task->id);
+        $assignedToLabelNames = [];
+        foreach ($taskCheck->labels as $labelName) {
+            $assignedToLabelNames[] = $labelName->name;
+        }
         $relationship = [
             'assignedToName' => $taskCheck->assignedTo->name ?? null,
-            'statusName' => $taskCheck->status->name,
+            'statusName' => $taskCheck->status->name ?? null,
+            'assignedToLabel' => $assignedToLabelNames,
         ];
-        return view('tasks.edit', compact('task', 'taskStatuses', 'users', 'relationship'));
+        return view('tasks.edit', compact('task', 'taskStatuses', 'users', 'labels', 'relationship'));
     }
 
     /**
@@ -106,15 +123,22 @@ class TaskController extends Controller
      */
     public function update(StoreTask $request, Task $task)
     {
-
+        $data = $request->input();
         $updatedTask = Task::find($task->id);
+        $updatedTask->labels()->detach($updatedTask->labels);
         $validated = $request->validated();
         $updatedTask->fill($validated);
         $updatedTask->save();
+        $newLabels = $data['labels'];
+        if ($newLabels[0] == null) {
+            unset($newLabels[0]);
+        }
+        if (count($newLabels) > 0) {
+            $updatedTask->labels()->attach($newLabels);
+        }
         flash(__('flash.tasks.update.success'))->success();
         return redirect()->route('tasks.index');
     }
-
     /**
      * Remove the specified resource from storage.
      *
@@ -123,11 +147,13 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
-        if ($task) {
+        $labels = $task->labels()->get();
+        if ($task && count($labels) == 0) {
             $task->delete();
+            flash(__('flash.tasks.delete.success'))->success();
+        } else {
+            flash(__('flash.tasks.failed_to_delete.error'))->error();
         }
-        flash(__('flash.tasks.delete.success'))->success();
-        // flash(__('flash.tasks.failed_to_delete.error'));
         return redirect()->route('tasks.index');
     }
 }
